// src/pages/search/attractions.interface.ts
export interface Card {
    name: string;
    desc: string;
    img: string;
    icon: string;
    stars: number;
    distance: number;
    price: number;
    seen: boolean;
    favourite: boolean;
    address: string;
    webpage: string;
    time: string;
    tags: string;
    GUI_isOpen: boolean;
}